third_party/blink/web_tests/editing/inserting/insert-space-at-start-of-wrapped-line.html

<!doctype HTML>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
#textarea, #plaintext, #richedit {
  font-family: 'Courier', monospace;
  width: 10ch;
}
#plaintext, #richedit {
  border: 1px solid black;
  white-space: pre-wrap;
}
</style>
<textarea id="textarea"></textarea>
<div id="plaintext" contenteditable="plaintext-only"></div>
<div id="richedit" contenteditable></div>
<script>
function runInElement(element, before, after) {
  test(() => {
    let offset = before.length;
    let acutal;
    if (element.tagName === 'TEXTAREA') {
      element.value = before + after;
      element.focus();
      element.setSelectionRange(offset, offset);
      eventSender.keyDown(' ');
      actual = element.value;
    } else {
      element.textContent = before + after;
      let textNode = element.firstChild;
      window.getSelection().setBaseAndExtent(textNode, offset, textNode, offset);
      eventSender.keyDown(' ');
      actual = element.textContent;
    }
    assert_equals(actual, before + ' ' + after);
  }, `Typing space at "${before}|${after}" should insert a space in ${element.id}`);
}

function run(before, after) {
  runInElement(textarea, before, after);
  runInElement(plaintext, before, after);
  runInElement(richedit, before, after);
}

(function () {
  if (!window.eventSender) {
    test(function () {
      assert_unreached();
    }, 'This test requires eventSender');
    return;
  }

  run('123456789 ', 'XY');
  run('12345     ', 'XY');
  run('12345        ', 'XY');
  run('123456789-', 'XY');
})();
</script>
